requirements.txt
websockets
google-cloud-speech
